[modbusmaster][reload config]: first prototype
