+++ extensions/ModbusMaster/MBTCPMultiMaster.cc
@@ -33,175 +33,168 @@ MBTCPMultiMaster::MBTCPMultiMaster( uniset::ObjectId objId, uniset::ObjectId shm
     MBExchange(objId, shmId, ic, prefix),
     force_disconnect(true)
 {
-    if( objId == DefaultObjectId )
-        throw uniset::SystemError("(MBTCPMultiMaster): objId=-1?!! Use --" + prefix + "-name" );
+	if( objId == DefaultObjectId )
+		throw uniset::SystemError("(MBTCPMultiMaster): objId=-1?!! Use --" + prefix + "-name" );
 
-    auto conf = uniset_conf();
+	auto conf = uniset_conf();
 
-    prop_prefix = initPropPrefix("tcp_");
-    mbinfo << myname << "(init): prop_prefix=" << prop_prefix << endl;
+	mbconf->prop_prefix = initPropPrefix(mbconf->s_field, "tcp_");
+	mbinfo << myname << "(init): prop_prefix=" << mbconf->prop_prefix << endl;
 
-    UniXML::iterator it(cnode);
+	UniXML::iterator it(cnode);
 
-    checktime = conf->getArgPInt("--" + prefix + "-checktime", it.getProp("checktime"), 5000);
-    force_disconnect = conf->getArgInt("--" + prefix + "-persistent-connection", it.getProp("persistent_connection")) ? false : true;
+	checktime = conf->getArgPInt("--" + prefix + "-checktime", it.getProp("checktime"), 5000);
+	force_disconnect = conf->getArgInt("--" + prefix + "-persistent-connection", it.getProp("persistent_connection")) ? false : true;
 
-    int ignore_timeout = conf->getArgPInt("--" + prefix + "-ignore-timeout", it.getProp("ignore_timeout"), ptReopen.getInterval());
+	int ignore_timeout = conf->getArgPInt("--" + prefix + "-ignore-timeout", it.getProp("ignore_timeout"), ptReopen.getInterval());
 
-    // Т.к. при "многоканальном" доступе к slave, смена канала должна происходит сразу после
-    // неудачной попытки запросов по одному из каналов, то ПЕРЕОПРЕДЕЛЯЕМ reopen, на channel-timeout..
-    int channelTimeout = conf->getArgPInt("--" + prefix + "-default-channel-timeout", it.getProp("channelTimeout"), default_timeout);
-    ptReopen.setTiming(channelTimeout);
+	// Т.к. при "многоканальном" доступе к slave, смена канала должна происходит сразу после
+	// неудачной попытки запросов по одному из каналов, то ПЕРЕОПРЕДЕЛЯЕМ reopen, на channel-timeout..
+	int channelTimeout = conf->getArgPInt("--" + prefix + "-default-channel-timeout", it.getProp("channelTimeout"), mbconf->default_timeout);
+	ptReopen.setTiming(channelTimeout);
 
-    UniXML::iterator it1(it);
+	UniXML::iterator it1(it);
 
-    if( !it1.find("GateList") )
-    {
-        ostringstream err;
-        err << myname << "(init): not found <GateList>";
-        mbcrit << err.str() << endl;
-        throw uniset::SystemError(err.str());
-    }
+	if( !it1.find("GateList") )
+	{
+		ostringstream err;
+		err << myname << "(init): not found <GateList>";
+		mbcrit << err.str() << endl;
+		throw uniset::SystemError(err.str());
+	}
 
-    if( !it1.goChildren() )
-    {
-        ostringstream err;
-        err << myname << "(init): empty <GateList> ?!";
-        mbcrit << err.str() << endl;
-        throw uniset::SystemError(err.str());
-    }
+	if( !it1.goChildren() )
+	{
+		ostringstream err;
+		err << myname << "(init): empty <GateList> ?!";
+		mbcrit << err.str() << endl;
+		throw uniset::SystemError(err.str());
+	}
 
-    for( ; it1.getCurrent(); it1++ )
-    {
-        if( it1.getIntProp("ignore") )
-        {
-            mbinfo << myname << "(init): IGNORE " <<  it1.getProp("ip") << ":" << it1.getProp("port") << endl;
-            continue;
-        }
+	for( ; it1.getCurrent(); it1++ )
+	{
+		if( it1.getIntProp("ignore") )
+		{
+			mbinfo << myname << "(init): IGNORE " <<  it1.getProp("ip") << ":" << it1.getProp("port") << endl;
+			continue;
+		}
 
-        auto sinf = make_shared<MBSlaveInfo>();
+		auto sinf = make_shared<MBSlaveInfo>();
 
-        sinf->ip = it1.getProp("ip");
+		sinf->ip = it1.getProp("ip");
 
-        if( sinf->ip.empty() )
-        {
-            ostringstream err;
-            err << myname << "(init): ip='' in <GateList>";
-            mbcrit << err.str() << endl;
-            throw uniset::SystemError(err.str());
-        }
+		if( sinf->ip.empty() )
+		{
+			ostringstream err;
+			err << myname << "(init): ip='' in <GateList>";
+			mbcrit << err.str() << endl;
+			throw uniset::SystemError(err.str());
+		}
 
-        sinf->port = it1.getIntProp("port");
+		sinf->port = it1.getIntProp("port");
 
-        if( sinf->port <= 0 )
-        {
-            ostringstream err;
-            err << myname << "(init): ERROR: port=''" << sinf->port << " for ip='" << sinf->ip << "' in <GateList>";
-            mbcrit << err.str() << endl;
-            throw uniset::SystemError(err.str());
-        }
+		if( sinf->port <= 0 )
+		{
+			ostringstream err;
+			err << myname << "(init): ERROR: port=''" << sinf->port << " for ip='" << sinf->ip << "' in <GateList>";
+			mbcrit << err.str() << endl;
+			throw uniset::SystemError(err.str());
+		}
 
-        if( !it1.getProp("respondSensor").empty() )
-        {
-            sinf->respond_id = conf->getSensorID( it1.getProp("respondSensor") );
+		if( !it1.getProp("respondSensor").empty() )
+		{
+			sinf->respond_id = conf->getSensorID( it1.getProp("respondSensor") );
 
-            if( sinf->respond_id == DefaultObjectId )
-            {
-                ostringstream err;
-                err << myname << "(init): ERROR: Unknown SensorID for '" << it1.getProp("respondSensor") << "' in <GateList>";
-                mbcrit << err.str() << endl;
-                throw uniset::SystemError(err.str());
-            }
-        }
+			if( sinf->respond_id == DefaultObjectId )
+			{
+				ostringstream err;
+				err << myname << "(init): ERROR: Unknown SensorID for '" << it1.getProp("respondSensor") << "' in <GateList>";
+				mbcrit << err.str() << endl;
+				throw uniset::SystemError(err.str());
+			}
+		}
 
-        sinf->priority = it1.getIntProp("priority");
-        sinf->mbtcp = std::make_shared<ModbusTCPMaster>();
+		sinf->priority = it1.getIntProp("priority");
+		sinf->mbtcp = std::make_shared<ModbusTCPMaster>();
 
-        sinf->ptIgnoreTimeout.setTiming( it1.getPIntProp("ignore_timeout", ignore_timeout) );
-        sinf->recv_timeout = it1.getPIntProp("recv_timeout", recv_timeout);
-        sinf->aftersend_pause = it1.getPIntProp("aftersend_pause", aftersend_pause);
-        sinf->sleepPause_usec = it1.getPIntProp("sleepPause_msec", sleepPause_msec);
-        sinf->respond_invert = it1.getPIntProp("invert", 0);
-        sinf->respond_force = it1.getPIntProp("force", 0);
+		sinf->ptIgnoreTimeout.setTiming( it1.getPIntProp("ignore_timeout", ignore_timeout) );
+		sinf->recv_timeout = it1.getPIntProp("recv_timeout", mbconf->recv_timeout);
+		sinf->aftersend_pause = it1.getPIntProp("aftersend_pause", mbconf->aftersend_pause);
+		sinf->sleepPause_usec = it1.getPIntProp("sleepPause_msec", mbconf->sleepPause_msec);
+		sinf->respond_invert = it1.getPIntProp("invert", 0);
+		sinf->respond_force = it1.getPIntProp("force", 0);
 
-        int fn = conf->getArgPInt("--" + prefix + "-check-func", it.getProp("checkFunc"), ModbusRTU::fnUnknown);
+		int fn = conf->getArgPInt("--" + prefix + "-check-func", it.getProp("checkFunc"), ModbusRTU::fnUnknown);
 
-        if( fn != ModbusRTU::fnUnknown &&
-                fn != ModbusRTU::fnReadCoilStatus &&
-                fn != ModbusRTU::fnReadInputStatus &&
-                fn != ModbusRTU::fnReadOutputRegisters &&
-                fn != ModbusRTU::fnReadInputRegisters )
-        {
-            ostringstream err;
-            err << myname << "(init):  BAD check function ='" << fn << "'. Must be [1,2,3,4]";
-            mbcrit << err.str() << endl;
-            throw SystemError(err.str());
-        }
+		if( fn != ModbusRTU::fnUnknown &&
+				fn != ModbusRTU::fnReadCoilStatus &&
+				fn != ModbusRTU::fnReadInputStatus &&
+				fn != ModbusRTU::fnReadOutputRegisters &&
+				fn != ModbusRTU::fnReadInputRegisters )
+		{
+			ostringstream err;
+			err << myname << "(init):  BAD check function ='" << fn << "'. Must be [1,2,3,4]";
+			mbcrit << err.str() << endl;
+			throw SystemError(err.str());
+		}
 
-        sinf->checkFunc = (ModbusRTU::SlaveFunctionCode)fn;
+		sinf->checkFunc = (ModbusRTU::SlaveFunctionCode)fn;
 
-        sinf->checkAddr = conf->getArgPInt("--" + prefix + "-check-addr", it.getProp("checkAddr"), 0);
-        sinf->checkReg = conf->getArgPInt("--" + prefix + "-check-reg", it.getProp("checkReg"), 0);
+		sinf->checkAddr = conf->getArgPInt("--" + prefix + "-check-addr", it.getProp("checkAddr"), 0);
+		sinf->checkReg = conf->getArgPInt("--" + prefix + "-check-reg", it.getProp("checkReg"), 0);
 
-        int tout = it1.getPIntProp("timeout", channelTimeout);
-        sinf->channel_timeout = (tout >= 0 ? tout : channelTimeout);
+		int tout = it1.getPIntProp("timeout", channelTimeout);
+		sinf->channel_timeout = (tout >= 0 ? tout : channelTimeout);
 
-        // делаем только задержку на отпускание..
-        sinf->respondDelay.set(0, sinf->channel_timeout);
+		// делаем только задержку на отпускание..
+		sinf->respondDelay.set(0, sinf->channel_timeout);
 
-        sinf->force_disconnect = it.getPIntProp("persistent_connection", !force_disconnect) ? false : true;
+		sinf->force_disconnect = it.getPIntProp("persistent_connection", !force_disconnect) ? false : true;
 
-        ostringstream n;
-        n << sinf->ip << ":" << sinf->port;
-        sinf->myname = n.str();
+		ostringstream n;
+		n << sinf->ip << ":" << sinf->port;
+		sinf->myname = n.str();
 
-        auto l = loga->create(sinf->myname);
-        sinf->mbtcp->setLog(l);
+		auto l = loga->create(sinf->myname);
+		sinf->mbtcp->setLog(l);
 
-        mbinfo << myname << "(init): add slave channel " << sinf->myname << endl;
-        mblist.emplace_back(sinf);
-    }
+		mbinfo << myname << "(init): add slave channel " << sinf->myname << endl;
+		mblist.emplace_back(sinf);
+	}
 
-    if( ic )
-        ic->logAgregator()->add(loga);
+	if( ic )
+		ic->logAgregator()->add(loga);
 
 
-    if( mblist.empty() )
-    {
-        ostringstream err;
-        err << myname << "(init): empty <GateList>!";
-        mbcrit << err.str() << endl;
-        throw uniset::SystemError(err.str());
-    }
+	if( mblist.empty() )
+	{
+		ostringstream err;
+		err << myname << "(init): empty <GateList>!";
+		mbcrit << err.str() << endl;
+		throw uniset::SystemError(err.str());
+	}
 
-    mblist.sort();
-    mbi = mblist.rbegin(); // т.к. mbi это reverse_iterator
-    (*mbi)->setUse(true);
+	mblist.sort();
+	mbi = mblist.rbegin(); // т.к. mbi это reverse_iterator
+	(*mbi)->setUse(true);
 
-    if( shm->isLocalwork() )
-    {
-        readConfiguration();
+	if( shm->isLocalwork() )
+		mbconf->loadConfig(conf->getConfXML(), conf->getXMLSensorsSection());
+	else
+		ic->addReadItem( sigc::mem_fun(this, &MBTCPMultiMaster::readItem) );
 
-        if( !noQueryOptimization )
-            rtuQueryOptimization(devices);
+	pollThread = unisetstd::make_unique<ThreadCreator<MBTCPMultiMaster>>(this, &MBTCPMultiMaster::poll_thread);
+	pollThread->setFinalAction(this, &MBTCPMultiMaster::final_thread);
+	checkThread = unisetstd::make_unique<ThreadCreator<MBTCPMultiMaster>>(this, &MBTCPMultiMaster::check_thread);
+	checkThread->setFinalAction(this, &MBTCPMultiMaster::final_thread);
 
-        initDeviceList();
-    }
-    else
-        ic->addReadItem( sigc::mem_fun(this, &MBTCPMultiMaster::readItem) );
+	// Т.к. при "многоканальном" доступе к slave, смена канала должна происходит сразу после
+	// неудачной попытки запросов по одному из каналов, то ПЕРЕОПРЕДЕЛЯЕМ reopen, на channel-timeout..
+	int tout = conf->getArgPInt("--" + prefix + "-default-channel-timeout", it.getProp("channelTimeout"), mbconf->default_timeout);
+	ptReopen.setTiming(tout);
 
-    pollThread = unisetstd::make_unique<ThreadCreator<MBTCPMultiMaster>>(this, &MBTCPMultiMaster::poll_thread);
-    pollThread->setFinalAction(this, &MBTCPMultiMaster::final_thread);
-    checkThread = unisetstd::make_unique<ThreadCreator<MBTCPMultiMaster>>(this, &MBTCPMultiMaster::check_thread);
-    checkThread->setFinalAction(this, &MBTCPMultiMaster::final_thread);
-
-    // Т.к. при "многоканальном" доступе к slave, смена канала должна происходит сразу после
-    // неудачной попытки запросов по одному из каналов, то ПЕРЕОПРЕДЕЛЯЕМ reopen, на channel-timeout..
-    int tout = conf->getArgPInt("--" + prefix + "-default-channel-timeout", it.getProp("channelTimeout"), default_timeout);
-    ptReopen.setTiming(tout);
-
-    if( mblog->is_info() )
-        printMap(devices);
+	if( mblog->is_info() )
+		MBConfig::printMap(mbconf->devices);
 }
 // -----------------------------------------------------------------------------
 MBTCPMultiMaster::~MBTCPMultiMaster()
@@ -233,111 +226,111 @@ MBTCPMultiMaster::~MBTCPMultiMaster()
 // -----------------------------------------------------------------------------
 std::shared_ptr<ModbusClient> MBTCPMultiMaster::initMB( bool reopen )
 {
-    if( mb )
-        ptInitChannel.reset();
+	if( mb )
+		ptInitChannel.reset();
 
-    // просто движемся по кругу (т.к. связь не проверяется)
-    // движемся в обратном порядке, т.к. сортировка по возрастанию приоритета
-    if( checktime <= 0 )
-    {
-        ++mbi;
+	// просто движемся по кругу (т.к. связь не проверяется)
+	// движемся в обратном порядке, т.к. сортировка по возрастанию приоритета
+	if( checktime <= 0 )
+	{
+		++mbi;
 
-        if( mbi == mblist.rend() )
-            mbi = mblist.rbegin();
+		if( mbi == mblist.rend() )
+			mbi = mblist.rbegin();
 
-        auto m = (*mbi);
-        m->init(mblog);
+		auto m = (*mbi);
+		m->init(mblog);
 
-        // переопределяем timeout на данный канал
-        ptReopen.setTiming( m->channel_timeout );
+		// переопределяем timeout на данный канал
+		ptReopen.setTiming( m->channel_timeout );
 
-        m->setUse(true);
-        mb = m->mbtcp;
-        return m->mbtcp;
-    }
+		m->setUse(true);
+		mb = m->mbtcp;
+		return m->mbtcp;
+	}
 
-    {
-        // сперва надо обновить все ignore
-        // т.к. фактически флаги выставляются и сбрасываются только здесь
-        for( auto&& it : mblist )
-            it->ignore = !it->ptIgnoreTimeout.checkTime();
+	{
+		// сперва надо обновить все ignore
+		// т.к. фактически флаги выставляются и сбрасываются только здесь
+		for( auto&& it : mblist )
+			it->ignore = !it->ptIgnoreTimeout.checkTime();
 
-        // если reopen=true - значит почему-то по текущему каналу связи нет (хотя соединение есть)
-        // тогда выставляем ему признак игнорирования
-        if( mbi != mblist.rend() && reopen )
-        {
-            auto m = (*mbi);
-            m->setUse(false);
-            m->ignore = true;
-            m->ptIgnoreTimeout.reset();
-            mbwarn << myname << "(initMB): set ignore=true for " << m->ip << ":" << m->port << endl;
-        }
+		// если reopen=true - значит почему-то по текущему каналу связи нет (хотя соединение есть)
+		// тогда выставляем ему признак игнорирования
+		if( mbi != mblist.rend() && reopen )
+		{
+			auto m = (*mbi);
+			m->setUse(false);
+			m->ignore = true;
+			m->ptIgnoreTimeout.reset();
+			mbwarn << myname << "(initMB): set ignore=true for " << m->ip << ":" << m->port << endl;
+		}
 
-        // Если по текущему каналу связь есть (и мы его не игнорируем), то возвращаем его
-        if( mbi != mblist.rend() && !(*mbi)->ignore && (*mbi)->respond )
-        {
-            auto m = (*mbi);
-            // ещё раз проверим соединение (в неблокирующем режиме)
-            m->respond = m->check();
+		// Если по текущему каналу связь есть (и мы его не игнорируем), то возвращаем его
+		if( mbi != mblist.rend() && !(*mbi)->ignore && (*mbi)->respond )
+		{
+			auto m = (*mbi);
+			// ещё раз проверим соединение (в неблокирующем режиме)
+			m->respond = m->check();
 
-            if( m->respond && (m->mbtcp->isConnection() || m->init(mblog)) )
-            {
-                mblog4 << myname << "(initMB): SELECT CHANNEL " << m->ip << ":" << m->port << endl;
-                mb = m->mbtcp;
-                m->setUse(true);
-                ptReopen.setTiming( m->channel_timeout );
-                return m->mbtcp;
-            }
+			if( m->respond && (m->mbtcp->isConnection() || m->init(mblog)) )
+			{
+				mblog4 << myname << "(initMB): SELECT CHANNEL " << m->ip << ":" << m->port << endl;
+				mb = m->mbtcp;
+				m->setUse(true);
+				ptReopen.setTiming( m->channel_timeout );
+				return m->mbtcp;
+			}
 
-            m->setUse(false);
-        }
+			m->setUse(false);
+		}
 
-        if( mbi != mblist.rend() )
-            (*mbi)->mbtcp->forceDisconnect();
-    }
+		if( mbi != mblist.rend() )
+			(*mbi)->mbtcp->forceDisconnect();
+	}
 
-    // проходим по списку (в обратном порядке, т.к. самый приоритетный в конце)
-    for( auto it = mblist.rbegin(); it != mblist.rend(); ++it )
-    {
-        auto m = (*it);
+	// проходим по списку (в обратном порядке, т.к. самый приоритетный в конце)
+	for( auto it = mblist.rbegin(); it != mblist.rend(); ++it )
+	{
+		auto m = (*it);
 
-        if( m->respond && !m->ignore && m->init(mblog) )
-        {
-            mbi = it;
-            mb = m->mbtcp;
-            m->setUse(true);
-            ptReopen.setTiming( m->channel_timeout );
-            mblog4 << myname << "(initMB): SELECT CHANNEL " << m->ip << ":" << m->port << endl;
-            return m->mbtcp;
-        }
-    }
+		if( m->respond && !m->ignore && m->init(mblog) )
+		{
+			mbi = it;
+			mb = m->mbtcp;
+			m->setUse(true);
+			ptReopen.setTiming( m->channel_timeout );
+			mblog4 << myname << "(initMB): SELECT CHANNEL " << m->ip << ":" << m->port << endl;
+			return m->mbtcp;
+		}
+	}
 
-    // если дошли сюда.. значит не нашли ни одного канала..
-    // но т.к. мы пропускали те, которые в ignore
-    // значит сейчас просто находим первый у кого есть связь и делаем его главным
-    for( auto it = mblist.rbegin(); it != mblist.rend(); ++it )
-    {
-        auto& m = (*it);
+	// если дошли сюда.. значит не нашли ни одного канала..
+	// но т.к. мы пропускали те, которые в ignore
+	// значит сейчас просто находим первый у кого есть связь и делаем его главным
+	for( auto it = mblist.rbegin(); it != mblist.rend(); ++it )
+	{
+		auto& m = (*it);
 
-        if( m->respond && m->check() && m->init(mblog) )
-        {
-            mbi = it;
-            mb = m->mbtcp;
-            m->ignore = false;
-            m->setUse(true);
-            ptReopen.setTiming( m->channel_timeout );
-            mblog4 << myname << "(initMB): SELECT CHANNEL " << m->ip << ":" << m->port << endl;
-            return m->mbtcp;
-        }
-    }
+		if( m->respond && m->check() && m->init(mblog) )
+		{
+			mbi = it;
+			mb = m->mbtcp;
+			m->ignore = false;
+			m->setUse(true);
+			ptReopen.setTiming( m->channel_timeout );
+			mblog4 << myname << "(initMB): SELECT CHANNEL " << m->ip << ":" << m->port << endl;
+			return m->mbtcp;
+		}
+	}
 
-    // значит всё-таки связи реально нет...
-    {
-        mbi = mblist.rend();
-        mb = nullptr;
-    }
+	// значит всё-таки связи реально нет...
+	{
+		mbi = mblist.rend();
+		mb = nullptr;
+	}
 
-    return 0;
+	return 0;
 }
 // -----------------------------------------------------------------------------
 void MBTCPMultiMaster::final_thread()
@@ -462,121 +455,120 @@ void MBTCPMultiMaster::sysCommand( const uniset::SystemMessage* sm )
 // -----------------------------------------------------------------------------
 void MBTCPMultiMaster::poll_thread()
 {
-    // ждём начала работы..(см. MBExchange::activateObject)
-    while( !isProcActive() && !canceled )
-    {
-        uniset::uniset_rwmutex_rlock l(mutex_start);
-    }
+	// ждём начала работы..(см. MBExchange::activateObject)
+	while( !isProcActive() && !canceled )
+	{
+		uniset::uniset_rwmutex_rlock l(mutex_start);
+	}
 
-    // работаем..
-    while( isProcActive() )
-    {
-        try
-        {
-            if( sidExchangeMode != DefaultObjectId && force )
-                exchangeMode = shm->localGetValue(itExchangeMode, sidExchangeMode);
-        }
-        catch( std::exception& ex )
-        {
-            mbwarn << myname << "(poll_thread): "  << ex.what() << endl;
-        }
+	// работаем..
+	while( isProcActive() )
+	{
+		try
+		{
+			if( sidExchangeMode != DefaultObjectId && force )
+				exchangeMode = shm->localGetValue(itExchangeMode, sidExchangeMode);
+		}
+		catch( std::exception& ex )
+		{
+			mbwarn << myname << "(poll_thread): "  << ex.what() << endl;
+		}
 
-        try
-        {
-            poll();
-        }
-        catch( std::exception& ex)
-        {
-            mbwarn << myname << "(poll_thread): "  << ex.what() << endl;
-        }
+		try
+		{
+			poll();
+		}
+		catch( std::exception& ex)
+		{
+			mbwarn << myname << "(poll_thread): "  << ex.what() << endl;
+		}
 
-        if( !isProcActive() )
-            break;
+		if( !isProcActive() )
+			break;
 
-        msleep(polltime);
-    }
+		msleep(mbconf->polltime);
+	}
 }
 // -----------------------------------------------------------------------------
 void MBTCPMultiMaster::check_thread()
 {
-    while( isProcActive() )
-    {
-        for( auto&& it : mblist )
-        {
-            try
-            {
-                // сбрасываем флаг ignore..раз время вышло.
-                it->ignore = !it->ptIgnoreTimeout.checkTime();
+	while( isProcActive() )
+	{
+		for( auto&& it : mblist )
+		{
+			try
+			{
+				// сбрасываем флаг ignore..раз время вышло.
+				it->ignore = !it->ptIgnoreTimeout.checkTime();
 
-                // Если use=1" связь не проверяем и считаем что связь есть..
-                bool r = ( it->use ? true : it->check() );
+				// Если use=1" связь не проверяем и считаем что связь есть..
+				bool r = ( it->use ? true : it->check() );
 
-                mblog4 << myname << "(check): " << it->myname << " " << setw(4) << ( r ? "OK" : "FAIL" )
-                       << " [ respondDelay=" << it->respondDelay.check( r )
-                       << " timeout=" << it->channel_timeout
-                       << " use=" << it->use
-                       << " ignore=" << it->ignore
-                       << " respond_id=" << it->respond_id
-                       << " respond_force=" << it->respond_force
-                       << " respond=" << it->respond
-                       << " respond_invert=" << it->respond_invert
-                       << " activated=" << isProcActive()
-                       << " ]"
-                       << endl;
+				mblog4 << myname << "(check): " << it->myname << " " << setw(4) << ( r ? "OK" : "FAIL" )
+					   << " [ respondDelay=" << it->respondDelay.check( r )
+					   << " timeout=" << it->channel_timeout
+					   << " use=" << it->use
+					   << " ignore=" << it->ignore
+					   << " respond_id=" << it->respond_id
+					   << " respond_force=" << it->respond_force
+					   << " respond=" << it->respond
+					   << " respond_invert=" << it->respond_invert
+					   << " activated=" << isProcActive()
+					   << " ]"
+					   << endl;
 
-                // задержка на выставление "пропажи связи"
-                if( it->respond_init )
-                    r = it->respondDelay.check( r );
+				// задержка на выставление "пропажи связи"
+				if( it->respond_init )
+					r = it->respondDelay.check( r );
 
-                if( !isProcActive() )
-                    break;
+				if( !isProcActive() )
+					break;
 
-                try
-                {
-                    if( it->respond_id != DefaultObjectId && (it->respond_force || !it->respond_init || r != it->respond) )
-                    {
-                        bool set = it->respond_invert ? !r : r;
-                        shm->localSetValue(it->respond_it, it->respond_id, (set ? 1 : 0), getId());
+				try
+				{
+					if( it->respond_id != DefaultObjectId && (it->respond_force || !it->respond_init || r != it->respond) )
+					{
+						bool set = it->respond_invert ? !r : r;
+						shm->localSetValue(it->respond_it, it->respond_id, (set ? 1 : 0), getId());
 
-                        {
-                            std::lock_guard<std::mutex> l(it->mutInit);
-                            it->respond_init = true;
-                        }
-                    }
-                }
-                catch( const uniset::Exception& ex )
-                {
-                    mbcrit << myname << "(check): (respond) " << it->myname << " : " << ex << std::endl;
-                }
-                catch( const std::exception& ex )
-                {
-                    mbcrit << myname << "(check): (respond) " << it->myname << " : " << ex.what() << std::endl;
-                }
+						{
+							std::lock_guard<std::mutex> l(it->mutInit);
+							it->respond_init = true;
+						}
+					}
+				}
+				catch( const uniset::Exception& ex )
+				{
+					mbcrit << myname << "(check): (respond) " << it->myname << " : " << ex << std::endl;
+				}
+				catch( const std::exception& ex )
+				{
+					mbcrit << myname << "(check): (respond) " << it->myname << " : " << ex.what() << std::endl;
+				}
 
-                it->respond = r;
-            }
-            catch( const std::exception& ex )
-            {
-                mbcrit << myname << "(check): (respond) "  << it->myname << " : " << ex.what() << std::endl;
-            }
+				it->respond = r;
+			}
+			catch( const std::exception& ex )
+			{
+				mbcrit << myname << "(check): (respond) "  << it->myname << " : " << ex.what() << std::endl;
+			}
 
-            if( !isProcActive() )
-                break;
-        }
+			if( !isProcActive() )
+				break;
+		}
 
-        if( !isProcActive() )
-            break;
+		if( !isProcActive() )
+			break;
 
-        msleep(checktime);
-    }
+		msleep(checktime);
+	}
 }
 // -----------------------------------------------------------------------------
 void MBTCPMultiMaster::initIterators()
 {
     MBExchange::initIterators();
-
-    for( auto&& it : mblist )
-        shm->initIterator(it->respond_it);
+	for( auto&& it : mblist )
+		shm->initIterator(it->respond_it);
 }
 // -----------------------------------------------------------------------------
 bool MBTCPMultiMaster::deactivateObject()
@@ -607,7 +599,7 @@ void MBTCPMultiMaster::initCheckConnectionParameters()
 {
     auto conf = uniset_conf();
 
-    bool initFromRegMap = ( findArgParam("--" + prefix + "-check-init-from-regmap", conf->getArgc(), conf->getArgv()) != -1 );
+	bool initFromRegMap = ( findArgParam("--" + mbconf->prefix + "-check-init-from-regmap", conf->getArgc(), conf->getArgv()) != -1 );
 
     if( !initFromRegMap )
         return;
@@ -623,18 +615,16 @@ void MBTCPMultiMaster::initCheckConnectionParameters()
     ModbusRTU::ModbusAddr checkAddr = { 0x00 };
     ModbusRTU::ModbusData checkReg = { 0 };
 
-    if( devices.empty() )
-    {
-
+	if( mbconf->devices.empty() )
+	{
         mbwarn << myname << "(init): devices list empty?!" << endl;
         return;
     }
 
-    // идём по устройствам
-    for( const auto& d : devices )
-    {
-        checkAddr = d.second->mbaddr;
-
+	// идём по устройствам
+	for( const auto& d : mbconf->devices )
+	{
+		checkAddr = d.second->mbaddr;
         if( d.second->pollmap.empty() )
             continue;
 
@@ -677,13 +667,13 @@ void MBTCPMultiMaster::initCheckConnectionParameters()
            << " mbreg=" << (int)checkReg << "(" << ModbusRTU::dat2str(checkReg) << ")"
            << endl;
 
-    // инициализируем..
-    for( auto&& m : mblist )
-    {
-        m->checkFunc = checkFunc;
-        m->checkAddr = checkAddr;
-        m->checkReg = checkReg;
-    }
+	// инициализируем..
+	for( auto&& m : mblist )
+	{
+		m->checkFunc = checkFunc;
+		m->checkAddr = checkAddr;
+		m->checkReg = checkReg;
+	}
 }
 
 // -----------------------------------------------------------------------------
