+++ extensions/ModbusMaster/start_fg_mbtcp.sh
@@ -16,6 +16,7 @@
 --mbtcp-persistent-connection 1 \
 --ulog-add-levels system \
 --mbtcp-run-logserver \
+--mbtcp-log-add-levels any \
 $*
 
 #--mbtcp-log-add-levels level4,level3 \
