+++ extensions/ModbusMaster/tests/MBTCPTestServer.h
@@ -10,133 +10,127 @@
 /*! Реализация MBTCPTestServer для тестирования */
 class MBTCPTestServer
 {
-    public:
-        MBTCPTestServer( const std::unordered_set<uniset::ModbusRTU::ModbusAddr>& vaddr, const std::string& inetaddr, int port = 502, bool verbose = false );
-        ~MBTCPTestServer();
+	public:
+		MBTCPTestServer( const std::unordered_set<uniset::ModbusRTU::ModbusAddr>& vaddr, const std::string& inetaddr, int port = 502, bool verbose = false );
+		~MBTCPTestServer();
 
-        inline void setVerbose( bool state )
-        {
-            verbose = state;
-        }
+		inline void setVerbose( bool state )
+		{
+			verbose = state;
+		}
 
-        inline void setReply( uint32_t val )
-        {
-            replyVal = val;
-        }
+		inline void setReply( uint32_t val )
+		{
+			replyVal = val;
+		}
 
-        inline void setIgnoreAddrMode( bool state )
-        {
-            if( sslot )
-                sslot->setIgnoreAddrMode(state);
-        }
+		void execute();    /*!< основной цикл работы */
+		void setLog( std::shared_ptr<DebugStream> dlog );
 
-        void execute();    /*!< основной цикл работы */
-        void setLog( std::shared_ptr<DebugStream> dlog );
+		inline bool isRunning()
+		{
+			return ( sslot && sslot->isActive() );
+		}
 
-        inline bool isRunning()
-        {
-            return ( sslot && sslot->isActive() );
-        }
+		inline void disableExchange( bool set = true )
+		{
+			disabled = set;
+		}
 
-        inline void disableExchange( bool set = true )
-        {
-            disabled = set;
-        }
+		inline bool getForceSingleCoilCmd()
+		{
+			return forceSingleCoilCmd;
+		}
+		inline int16_t getLastWriteOutputSingleRegister()
+		{
+			return lastWriteOutputSingleRegister;
+		}
+		inline uniset::ModbusRTU::ForceCoilsMessage getLastForceCoilsQ()
+		{
+			return lastForceCoilsQ;
+		}
+		inline uniset::ModbusRTU::WriteOutputMessage getLastWriteOutput()
+		{
+			return lastWriteOutputQ;
+		}
 
-        inline bool getForceSingleCoilCmd()
-        {
-            return forceSingleCoilCmd;
-        }
-        inline int16_t getLastWriteOutputSingleRegister()
-        {
-            return lastWriteOutputSingleRegister;
-        }
-        inline uniset::ModbusRTU::ForceCoilsMessage getLastForceCoilsQ()
-        {
-            return lastForceCoilsQ;
-        }
-        inline uniset::ModbusRTU::WriteOutputMessage getLastWriteOutput()
-        {
-            return lastWriteOutputQ;
-        }
+		friend std::ostream& operator<<(std::ostream& os, const MBTCPTestServer* m );
 
-        friend std::ostream& operator<<(std::ostream& os, const MBTCPTestServer* m );
+		inline float getF2TestValue()
+		{
+			return f2_test_value;
+		}
 
-        inline float getF2TestValue()
-        {
-            return f2_test_value;
-        }
+	protected:
+		// действия при завершении работы
+		void sigterm( int signo );
 
-    protected:
-        // действия при завершении работы
-        void sigterm( int signo );
+		/*! обработка 0x01 */
+		uniset::ModbusRTU::mbErrCode readCoilStatus( uniset::ModbusRTU::ReadCoilMessage& query,
+				uniset::ModbusRTU::ReadCoilRetMessage& reply );
+		/*! обработка 0x02 */
+		uniset::ModbusRTU::mbErrCode readInputStatus( uniset::ModbusRTU::ReadInputStatusMessage& query,
+				uniset::ModbusRTU::ReadInputStatusRetMessage& reply );
 
-        /*! обработка 0x01 */
-        uniset::ModbusRTU::mbErrCode readCoilStatus( uniset::ModbusRTU::ReadCoilMessage& query,
-                uniset::ModbusRTU::ReadCoilRetMessage& reply );
-        /*! обработка 0x02 */
-        uniset::ModbusRTU::mbErrCode readInputStatus( uniset::ModbusRTU::ReadInputStatusMessage& query,
-                uniset::ModbusRTU::ReadInputStatusRetMessage& reply );
+		/*! обработка 0x03 */
+		uniset::ModbusRTU::mbErrCode readOutputRegisters( uniset::ModbusRTU::ReadOutputMessage& query,
+				uniset::ModbusRTU::ReadOutputRetMessage& reply );
 
-        /*! обработка 0x03 */
-        uniset::ModbusRTU::mbErrCode readOutputRegisters( uniset::ModbusRTU::ReadOutputMessage& query,
-                uniset::ModbusRTU::ReadOutputRetMessage& reply );
+		/*! обработка 0x04 */
+		uniset::ModbusRTU::mbErrCode readInputRegisters( uniset::ModbusRTU::ReadInputMessage& query,
+				uniset::ModbusRTU::ReadInputRetMessage& reply );
 
-        /*! обработка 0x04 */
-        uniset::ModbusRTU::mbErrCode readInputRegisters( uniset::ModbusRTU::ReadInputMessage& query,
-                uniset::ModbusRTU::ReadInputRetMessage& reply );
+		/*! обработка 0x05 */
+		uniset::ModbusRTU::mbErrCode forceSingleCoil( uniset::ModbusRTU::ForceSingleCoilMessage& query,
+				uniset::ModbusRTU::ForceSingleCoilRetMessage& reply );
 
-        /*! обработка 0x05 */
-        uniset::ModbusRTU::mbErrCode forceSingleCoil( uniset::ModbusRTU::ForceSingleCoilMessage& query,
-                uniset::ModbusRTU::ForceSingleCoilRetMessage& reply );
-
-        /*! обработка 0x0F */
-        uniset::ModbusRTU::mbErrCode forceMultipleCoils( uniset::ModbusRTU::ForceCoilsMessage& query,
-                uniset::ModbusRTU::ForceCoilsRetMessage& reply );
+		/*! обработка 0x0F */
+		uniset::ModbusRTU::mbErrCode forceMultipleCoils( uniset::ModbusRTU::ForceCoilsMessage& query,
+				uniset::ModbusRTU::ForceCoilsRetMessage& reply );
 
 
-        /*! обработка 0x10 */
-        uniset::ModbusRTU::mbErrCode writeOutputRegisters( uniset::ModbusRTU::WriteOutputMessage& query,
-                uniset::ModbusRTU::WriteOutputRetMessage& reply );
+		/*! обработка 0x10 */
+		uniset::ModbusRTU::mbErrCode writeOutputRegisters( uniset::ModbusRTU::WriteOutputMessage& query,
+				uniset::ModbusRTU::WriteOutputRetMessage& reply );
 
-        /*! обработка 0x06 */
-        uniset::ModbusRTU::mbErrCode writeOutputSingleRegister( uniset::ModbusRTU::WriteSingleOutputMessage& query,
-                uniset::ModbusRTU::WriteSingleOutputRetMessage& reply );
-
-
-        uniset::ModbusRTU::mbErrCode diagnostics( uniset::ModbusRTU::DiagnosticMessage& query,
-                uniset::ModbusRTU::DiagnosticRetMessage& reply );
-
-        uniset::ModbusRTU::mbErrCode read4314( uniset::ModbusRTU::MEIMessageRDI& query,
-                                               uniset::ModbusRTU::MEIMessageRetRDI& reply );
-
-        /*! обработка запросов на чтение ошибок */
-        uniset::ModbusRTU::mbErrCode journalCommand( uniset::ModbusRTU::JournalCommandMessage& query,
-                uniset::ModbusRTU::JournalCommandRetMessage& reply );
-
-        /*! обработка запроса на установку времени */
-        uniset::ModbusRTU::mbErrCode setDateTime( uniset::ModbusRTU::SetDateTimeMessage& query,
-                uniset::ModbusRTU::SetDateTimeRetMessage& reply );
-
-        /*! обработка запроса удалённого сервиса */
-        uniset::ModbusRTU::mbErrCode remoteService( uniset::ModbusRTU::RemoteServiceMessage& query,
-                uniset::ModbusRTU::RemoteServiceRetMessage& reply );
-
-        uniset::ModbusRTU::mbErrCode fileTransfer( uniset::ModbusRTU::FileTransferMessage& query,
-                uniset::ModbusRTU::FileTransferRetMessage& reply );
+		/*! обработка 0x06 */
+		uniset::ModbusRTU::mbErrCode writeOutputSingleRegister( uniset::ModbusRTU::WriteSingleOutputMessage& query,
+				uniset::ModbusRTU::WriteSingleOutputRetMessage& reply );
 
 
-        /*! интерфейс ModbusSlave для обмена по RS */
-        uniset::ModbusTCPServerSlot* sslot;
-        std::unordered_set<uniset::ModbusRTU::ModbusAddr> vaddr; /*!< адреса данного узла */
+		uniset::ModbusRTU::mbErrCode diagnostics( uniset::ModbusRTU::DiagnosticMessage& query,
+				uniset::ModbusRTU::DiagnosticRetMessage& reply );
 
-        bool verbose;
-        uint32_t replyVal;
-        bool forceSingleCoilCmd;
-        int16_t lastWriteOutputSingleRegister;
-        uniset::ModbusRTU::ForceCoilsMessage lastForceCoilsQ;
-        uniset::ModbusRTU::WriteOutputMessage lastWriteOutputQ;
-        float f2_test_value = {0.0};
+		uniset::ModbusRTU::mbErrCode read4314( uniset::ModbusRTU::MEIMessageRDI& query,
+											   uniset::ModbusRTU::MEIMessageRetRDI& reply );
+
+		/*! обработка запросов на чтение ошибок */
+		uniset::ModbusRTU::mbErrCode journalCommand( uniset::ModbusRTU::JournalCommandMessage& query,
+				uniset::ModbusRTU::JournalCommandRetMessage& reply );
+
+		/*! обработка запроса на установку времени */
+		uniset::ModbusRTU::mbErrCode setDateTime( uniset::ModbusRTU::SetDateTimeMessage& query,
+				uniset::ModbusRTU::SetDateTimeRetMessage& reply );
+
+		/*! обработка запроса удалённого сервиса */
+		uniset::ModbusRTU::mbErrCode remoteService( uniset::ModbusRTU::RemoteServiceMessage& query,
+				uniset::ModbusRTU::RemoteServiceRetMessage& reply );
+
+		uniset::ModbusRTU::mbErrCode fileTransfer( uniset::ModbusRTU::FileTransferMessage& query,
+				uniset::ModbusRTU::FileTransferRetMessage& reply );
+
+
+		/*! интерфейс ModbusSlave для обмена по RS */
+		uniset::ModbusTCPServerSlot* sslot;
+		std::unordered_set<uniset::ModbusRTU::ModbusAddr> vaddr; /*!< адреса данного узла */
+
+		bool verbose;
+		uint32_t replyVal;
+		bool forceSingleCoilCmd;
+		int16_t lastWriteOutputSingleRegister;
+		uniset::ModbusRTU::ForceCoilsMessage lastForceCoilsQ;
+		uniset::ModbusRTU::WriteOutputMessage lastWriteOutputQ;
+		float f2_test_value = {0.0};
 
 #if 0
         typedef std::map<uniset::ModbusRTU::mbErrCode, unsigned int> ExchangeErrorMap;
