+++ extensions/ModbusMaster/tests/run_test_mbtcpmaster.sh
@@ -15,8 +15,8 @@ cd -
 --mbtcp-filter-value 1 \
 --mbtcp-gateway-iaddr localhost \
 --mbtcp-gateway-port 20048 \
---mbtcp-polltime 50 --mbtcp-recv-timeout 500 
-#--mbtcp-log-add-levels any
+--mbtcp-polltime 50 --mbtcp-recv-timeout 500
+# --mbtcp-log-add-levels any
 #--mbtcp-default-mbinit-ok 1
 #--dlog-add-levels any
